Add offset configuration option
Motivation: walking distance from my home to the bus stop is around
5 minutes. Sensor usually displays a departure that I won't be able to
catch.

This commit introduces a new parameter "offset" which is a positive
integer value in seconds.
On each updated of departures sensor will check availability of the
offset parameter and it will add it to the start_time.
In the CZPubTran library (not the sensor) there is no value attached
to the payload if start_time is None. It is one of the drawbacks
of the current implementation as sensors gets current time on
home assistant instance and there might be a delay while it will submit
the request and update the sensor. Possible workaround is include
sensor update time to the offset.

I've introduced a new sensor attribute to display start time with the
offset. It might be useful to display sensor data while it it searching
for connection and departure time attribute isn't yet available.